requirements.txt

streamlit==1.24.0
opencv-python==4.7.0.72
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
openpyxl==3.1.2
face-recognition==1.3.0
requests==2.31.0
